requirements.txt
opencv-contrib-python>=4.11.0.86
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
numpy==1.26.4
pandas==2.2.3
matplotlib>=3.6.0
pillow==11.1.0
customtkinter>=5.2.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
paddleocr==2.9.1
paddlepaddle==2.6.2
seaborn==0.13.2
cvzone>=1.5.4
CTkMessagebox
openpyxl